+++ config/HISTORY.md
@@ -1,3 +1,8 @@
+# Su, 19 November 2017 | v0.0.2
+
+- full name on field's name if embedded; `DBCredentials.Username` instead of just `Username`.
+- able to parse any custom `flag.FlagSet`, see the [new example](_examples/file-and-flags-and-survey).
+
 # Fr, 17 November 2017 | v0.0.1
 
 Initial working release.